renamed atomics folder header files names according to good coding conventions, updated relative paths in other files and README, capitalized the name of constants used in all header files
