+++ src/top_model/main.cpp
@@ -21,9 +21,9 @@
 
 #include "../../include/data_structures/message.hpp"
 
-#include "../../include/atomics/senderCadmium.hpp"
-#include "../../include/atomics/receiverCadmium.hpp"
-#include "../../include/atomics/subnetCadmium.hpp"
+#include "../../include/atomics/sender_cadmium.hpp"
+#include "../../include/atomics/receiver_cadmium.hpp"
+#include "../../include/atomics/subnet_cadmium.hpp"
 
 
 using namespace std;
