+++ test/src/main_receiver.cpp
@@ -17,7 +17,7 @@
 #include "../../lib/vendor/NDTime.hpp"
 #include "../../lib/vendor/iestream.hpp"
 #include "../../include/data_structures/message.hpp"
-#include "../../include/atomics/receiverCadmium.hpp"
+#include "../../include/atomics/receiver_cadmium.hpp"
 
 
 using namespace std;
