+++ test/src/main_subnet.cpp
@@ -18,7 +18,7 @@
 #include "../../lib/vendor/NDTime.hpp"
 #include "../../lib/vendor/iestream.hpp"
 #include "../../include/data_structures/message.hpp"
-#include "../../include/atomics/subnetCadmium.hpp"
+#include "../../include/atomics/subnet_cadmium.hpp"
 
 using namespace std;
